Bug/en 1338 defer is not executed (#84)
consensus/chronology package

Fixed bug which could cause a situation when not all subrounds would been executed in the current round. (if the node would been stuck for a few seconds, the updateRound method called in chronology could change the round index and set the current subround to startRound)
consensus/spos/bn package

bnFactory.go → Changed a little bit the subrounds duration inside the round to improve the
consensus and added/changed the threshold percent parameters for the same reason.

bnWorker.go

a) Fixed a bug in ProcessReceivedMessages method when early messages received for the next round, at the edge between two rounds, from nodes which are already advanced to n + 1, would been ignored if the receiver would be canceled the current round n from any reasons and it would still be in the round n at the moment when the message would been received.

b) Fixed a bug when processing asynchronous received blocks, which would cause some parallel writing in account state and computing wrong root hash. (now there is a flag named ProcessingBlock which will help the system to know when is safe to advance to the next round. This over protection comes alongside the havetime method which will help the block processor to know when the time is running out)

c) The processingThresholdPercent was decreased from 85% to 75% for better reaction of the ProcessBlock method from block processor.

d) RevertAccountState method is now called from extend method which will be always executed by each node, if the consensus for committing and broadcasting the proposed block will not be achieved for any reasons. Actually extend method will be always executed when subroundEndRound will not succeeded.

bnSubround.go → Fixed a bug in RemainingTime method, which cause that the value of the given parameter startTime, to be reevaluated to each call, as the struct used (time.Time), is given to this method as a reference to the original object.

bnSubroundBlock.go

a) Fixed a bug in RemainingTime method, which cause that the value of the given parameter startTime, to be reevaluated to each call, as the struct used (time.Time), is given to this method as a reference to the original object.

b) Added a flag named ProcessingBlock in processReceivedBlock method which will help the system to know when the processing mechanism has been finished.

bnSubroundStartRound.go → Fixed a bug in RemainingTime method, which cause that the value of the given parameter startTime, to be reevaluated to each call, as the struct used (time.Time), is given to this method as a reference to the original object.
data/block package

block.go → Fixed a bug in IntegrityAndValidity method, which cause that for every block which would not contain transactions inside, this method would been return error
process/block package

process.go
a) Fixed bugs in ProcessBlock and CommitBlock methods when RevertAccountState method from defer method would not been executed in some cases.

b) Improved process block speed by changing get with peek for quick access to the transaction pool and also launching all the calls of OnRequestTransaction method through go routines

process/sync package

basicForkDetector.go
a) Changed in ForkDetector interface in method AddHeader the name of parameter isReceived -> isProcessed to be more close of its meaning

b) Added checkpoint for received nonces of signed and processed blocks

c) Fixed a bug when a roll back after a fork detection, would have been removed all the headers stored at that specific nonce and not only the rolled back one.

block.go
a) Improved the ShoudSync method which would return the synchronized state of the node directly, without needing to recalculate it in the same round, if it already was set as synchronized.

b) Improved the SyncBlock method to first resolve the forks and then to syncronize.

c) Removed the call of RevertAccountState method from CreateAndCommitEmptyBlock method, as this action should be managed by the caller.

* Refactored after code review

* Fixed mock interface after merge